refactor: share summary card util and clean startup templates
