+++ src/calculators/suite/profitability/ProfitabilityTab.tsx
@@ -7,8 +7,9 @@ import { ImportExport } from '@/components/ui/UIComponents/ImportExport';
 import ProfitabilityForm from './ProfitabilityForm';
 import { businessTypes, type BusinessType } from '../../../data/businessTypes';
 import { allFeedbackRules, type FeedbackItem, type CalculatorType } from '../../../data/feedbackRules'; 
-import { generateFeedback } from '../../../utils/feedbackUtils'; 
-import { FeedbackDrawer } from '../../../components/feedback/FeedbackDrawer'; 
+import { generateFeedback } from '../../../utils/feedbackUtils';
+import { FeedbackDrawer } from '../../../components/feedback/FeedbackDrawer';
+import { getSummaryCardClasses } from '@/utils/ui/getSummaryCardClasses';
 import { Label } from '@/components/ui/label';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import ProfitabilityChart from './ProfitabilityChart';
@@ -101,25 +102,10 @@ export default function ProfitabilityTab() {
     else setCompletionPercentage(100);
   }, [metrics]);
 
-  const getSummaryCardClassNameForProfitability = (metricName: string, currentFeedbackItems: FeedbackItem[]): string => {
-    const relevantFeedback = currentFeedbackItems.find(
-      (item) => item.uiTarget?.scope === 'summaryMetric' && item.uiTarget?.identifier === metricName
-    );
-
-    if (relevantFeedback) {
-      switch (relevantFeedback.severity) {
-        case 'critical':
-          return 'border-l-4 border-red-500';
-        case 'warning':
-          return 'border-l-4 border-yellow-500';
-        case 'good': // Corrected from 'success'
-          return 'border-l-4 border-green-500';
-        default:
-          return '';
-      }
-    }
-    return '';
-  };
+  const getSummaryCardClassNameForProfitability = (
+    metricName: string,
+    currentFeedbackItems: FeedbackItem[]
+  ): string => getSummaryCardClasses(metricName, currentFeedbackItems);
 
   return (
     <>
