+++ src/calculators/suite/startupcost/StartupCostTab.tsx
@@ -20,18 +20,13 @@ import {
 import { ImportExport } from '@/components/ui/UIComponents/ImportExport';
 import GuideCard from '@/components/ui/guide-card';
 import { businessTypes, type BusinessType } from '@/data/businessTypes';
+import { startupCostTemplates } from '@/data/startupCostTemplates';
+import { type CostItem } from '@/types/cost';
+import { getSummaryCardClasses } from '@/utils/ui/getSummaryCardClasses';
 // import { Progress } from '@/components/ui/progress'; For future use
 import { allFeedbackRules, type FeedbackItem, type CalculatorType } from '../../../data/feedbackRules';
 import { generateFeedback } from '../../../utils/feedbackUtils';
 import { FeedbackDrawer } from '../../../components/feedback/FeedbackDrawer';
-
-interface CostItem {
-  id: string;
-  name: string;
-  amount: number;
-  category: string;
-  isOneTime: boolean;
-}
 
 export default function StartupCostTab() {
   const [businessType, setBusinessType] = useState<string>('retail');
@@ -52,22 +47,13 @@ export default function StartupCostTab() {
   ]);
 
 
-  const getSummaryCardClassName = (metricName: string, defaultClasses: string): string => {
-    const relevantFeedback = feedbackItems.find(
-      fb => fb.uiTarget?.scope === 'summaryMetric' && fb.uiTarget?.identifier === metricName
-    );
-    const severity = relevantFeedback?.severity;
-    if (severity) {
-      switch (severity) {
-        case 'critical': return 'p-4 border-l-4 border-destructive bg-destructive bg-opacity-10';
-        case 'warning': return 'p-4 border-l-4 border-yellow-400 bg-yellow-400 bg-opacity-10';
-        case 'info': return 'p-4 border-l-4 border-blue-400 bg-blue-400 bg-opacity-10';
-        case 'good': return 'p-4 border-l-4 border-green-400 bg-green-400 bg-opacity-10';
-        default: return defaultClasses;
-      }
-    }
-    return defaultClasses;
-  };
+  const getSummaryCardClassName = (metricName: string, defaultClasses: string): string =>
+    getSummaryCardClasses(metricName, feedbackItems, defaultClasses, {
+      critical: 'p-4 border-l-4 border-destructive bg-destructive bg-opacity-10',
+      warning: 'p-4 border-l-4 border-yellow-400 bg-yellow-400 bg-opacity-10',
+      info: 'p-4 border-l-4 border-blue-400 bg-blue-400 bg-opacity-10',
+      good: 'p-4 border-l-4 border-green-400 bg-green-400 bg-opacity-10',
+    });
 
   // Calculate completion percentage based on number of items with a non-zero amount
   useEffect(() => {
@@ -107,111 +93,7 @@ export default function StartupCostTab() {
   };
 
   const applyTemplate = (type: string) => {
-    let templateItems: CostItem[] = [];
-    
-    switch(type) {
-      case 'retail':
-        templateItems = [
-          { id: 'retail1', name: 'Business Registration', amount: 500, category: 'Legal', isOneTime: true },
-          { id: 'retail2', name: 'Store Equipment', amount: 15000, category: 'Equipment', isOneTime: true },
-          { id: 'retail3', name: 'Rent Deposit', amount: 5000, category: 'Facilities', isOneTime: true },
-          { id: 'retail4', name: 'Initial Inventory', amount: 25000, category: 'Inventory', isOneTime: true },
-          { id: 'retail5', name: 'Store Fixtures', amount: 8000, category: 'Equipment', isOneTime: true },
-          { id: 'retail6', name: 'Point of Sale System', amount: 3000, category: 'Equipment', isOneTime: true },
-          { id: 'retail7', name: 'Signage', amount: 2500, category: 'Marketing', isOneTime: true },
-          { id: 'retail8', name: 'Website', amount: 2000, category: 'Marketing', isOneTime: true },
-          { id: 'retail9', name: 'Monthly Rent', amount: 2500, category: 'Facilities', isOneTime: false },
-          { id: 'retail10', name: 'Utilities', amount: 500, category: 'Facilities', isOneTime: false },
-          { id: 'retail11', name: 'Insurance', amount: 400, category: 'Insurance', isOneTime: false },
-          { id: 'retail12', name: 'Marketing', amount: 1500, category: 'Marketing', isOneTime: false },
-          { id: 'retail13', name: 'Employee Salaries', amount: 8000, category: 'Staffing', isOneTime: false },
-        ];
-        break;
-      case 'service':
-        templateItems = [
-          { id: 'service1', name: 'Business Registration', amount: 500, category: 'Legal', isOneTime: true },
-          { id: 'service2', name: 'Office Equipment', amount: 5000, category: 'Equipment', isOneTime: true },
-          { id: 'service3', name: 'Security Deposit', amount: 2000, category: 'Facilities', isOneTime: true },
-          { id: 'service4', name: 'Software & Licenses', amount: 3000, category: 'Equipment', isOneTime: true },
-          { id: 'service5', name: 'Website', amount: 3000, category: 'Marketing', isOneTime: true },
-          { id: 'service6', name: 'Professional Certifications', amount: 1500, category: 'Legal', isOneTime: true },
-          { id: 'service7', name: 'Monthly Rent', amount: 1500, category: 'Facilities', isOneTime: false },
-          { id: 'service8', name: 'Utilities', amount: 300, category: 'Facilities', isOneTime: false },
-          { id: 'service9', name: 'Insurance', amount: 350, category: 'Insurance', isOneTime: false },
-          { id: 'service10', name: 'Marketing', amount: 1000, category: 'Marketing', isOneTime: false },
-          { id: 'service11', name: 'Employee Salaries', amount: 6000, category: 'Staffing', isOneTime: false },
-        ];
-        break;
-      case 'online':
-        templateItems = [
-          { id: 'online1', name: 'Business Registration', amount: 500, category: 'Legal', isOneTime: true },
-          { id: 'online2', name: 'Website Development', amount: 5000, category: 'Marketing', isOneTime: true },
-          { id: 'online3', name: 'Logo & Branding', amount: 1500, category: 'Marketing', isOneTime: true },
-          { id: 'online4', name: 'Computer Equipment', amount: 3000, category: 'Equipment', isOneTime: true },
-          { id: 'online5', name: 'Initial Inventory', amount: 10000, category: 'Inventory', isOneTime: true },
-          { id: 'online6', name: 'E-commerce Platform', amount: 2000, category: 'Equipment', isOneTime: true },
-          { id: 'online7', name: 'Website Hosting', amount: 30, category: 'Equipment', isOneTime: false },
-          { id: 'online8', name: 'Online Marketing', amount: 1500, category: 'Marketing', isOneTime: false },
-          { id: 'online9', name: 'Software Subscriptions', amount: 200, category: 'Equipment', isOneTime: false },
-          { id: 'online10', name: 'Shipping Supplies', amount: 300, category: 'Inventory', isOneTime: false },
-          { id: 'online11', name: 'Insurance', amount: 200, category: 'Insurance', isOneTime: false },
-        ];
-        break;
-      case 'restaurant':
-        templateItems = [
-          { id: 'restaurant1', name: 'Business Registration', amount: 1000, category: 'Legal', isOneTime: true },
-          { id: 'restaurant2', name: 'Kitchen Equipment', amount: 50000, category: 'Equipment', isOneTime: true },
-          { id: 'restaurant3', name: 'Furniture & Fixtures', amount: 30000, category: 'Equipment', isOneTime: true },
-          { id: 'restaurant4', name: 'Rent Deposit', amount: 10000, category: 'Facilities', isOneTime: true },
-          { id: 'restaurant5', name: 'Renovations', amount: 40000, category: 'Facilities', isOneTime: true },
-          { id: 'restaurant6', name: 'Initial Inventory', amount: 15000, category: 'Inventory', isOneTime: true },
-          { id: 'restaurant7', name: 'Liquor License', amount: 5000, category: 'Legal', isOneTime: true },
-          { id: 'restaurant8', name: 'POS System', amount: 5000, category: 'Equipment', isOneTime: true },
-          { id: 'restaurant9', name: 'Monthly Rent', amount: 5000, category: 'Facilities', isOneTime: false },
-          { id: 'restaurant10', name: 'Utilities', amount: 1500, category: 'Facilities', isOneTime: false },
-          { id: 'restaurant11', name: 'Insurance', amount: 1000, category: 'Insurance', isOneTime: false },
-          { id: 'restaurant12', name: 'Marketing', amount: 3000, category: 'Marketing', isOneTime: false },
-          { id: 'restaurant13', name: 'Staff Salaries', amount: 20000, category: 'Staffing', isOneTime: false },
-          { id: 'restaurant14', name: 'Food & Beverage', amount: 10000, category: 'Inventory', isOneTime: false },
-        ];
-        break;
-      case 'saas':
-        templateItems = []; // Placeholder for SaaS template
-        break;
-      case 'manufacturing':
-        templateItems = []; // Placeholder for Manufacturing template
-        break;
-      case 'ecommerce':
-        templateItems = []; // Placeholder for E-commerce template
-        break;
-      case 'boutique':
-        templateItems = []; // Placeholder for Boutique template
-        break;
-      case 'coworking':
-        templateItems = []; // Placeholder for Co-working Space template
-        break;
-      case 'consulting':
-        templateItems = []; // Placeholder for Consulting Agency template
-        break;
-      case 'creative':
-        templateItems = []; // Placeholder for Creative Studio template
-        break;
-      case 'digitalMarketing':
-        templateItems = []; // Placeholder for Digital Marketing Agency template
-        break;
-      case 'eventPlanning':
-        templateItems = []; // Placeholder for Event Planning template
-        break;
-      case 'foodTruck':
-        templateItems = []; // Placeholder for Food Truck template
-        break;
-      case 'health':
-        templateItems = []; // Placeholder for Health & Wellness template
-        break;
-      default:
-        templateItems = items;
-    }
-    
+    const templateItems = startupCostTemplates[type] ?? items;
     setItems(templateItems);
     setBusinessType(type);
   };
