+++ src/types/cost.ts
@@ -1,0 +1,7 @@
+export interface CostItem {
+  id: string;
+  name: string;
+  amount: number;
+  category: string;
+  isOneTime: boolean;
+}
